Break apart into smaller modules
